Separate definition and implementation of IntegerConstant
